Add CoinSpawner class
CoinSpawner currently has 3 methods:
InstantiateCoinPrefabPool, InstantiateCoinPrefab, and
GetPooledCoin.

InstantiateCoinPrefabPool creates an initial pool of
coin prefabs that can be used when spawning coins onto
the field.

InstantiateCoinPrefab InstantiateCoinPrefab instantiates
a coin prefab and adds it to pooledObjsList. It also sets
the active state of the prefab's collider and sprite
renderer.

GetPooledCoin returns an inactive pooled coin.