Modify CoinSpawner and CoinBehaviour
Commented out Start method in CoinBehaviour.

Added SpawnCoinOnField method. This method selects and
inactive coin from the coin pool. Then it sets up the
coin value and enables the collider and sprite renderer.
Then it calls SelectValidSpawnCoord selects a valid spawn
location.

At the moment, there is an issue with SelectValidSpawnCoord
where it is allowing coordinates that correspond to the coin hole
to be used as a spawn location. It should not be doing this.