Add Score calculation functionality
PlayerStorage has a new method, CalcCoinScore, that
calculates the score when the player deposits coins.

GameProperties has been modified to load all
CoinScriptable objects. These are used in
CoinSpawner and PlayerStorage.

ScoreTextProperties subscribes UpdateScore to
PlayerStorage.DepositCoin.

When DepositCoin is invoked, it passes back the
calculated value of the deposited coins plus a
bonus based on the number of coin types.

The score text is then updated.