Modify UIManager and UIs object
UIManager has new methods for siwtching the "screens" on
the title menu: SetTitleGameObjectRefs and
SwitchMainTitleScreen.

SetTitleGameObjectRefs finds the relevant game objects
that contains the parent object related to the credit and
instruction screens. Then it sets the objects to variables.

SwitchMainTitleScreen enables credit and instruction
objects based on a boolean variable, _isCreditMenuEnabled.

The credit and instruction buttons in the UIs gameobject
now calls SwitchMainTitleScreen onclick.