Modify UIManager and PlayerControl
PlayerControl has a new delegate, TogglePauseUI.
This is invoked when the "pause" key is pressed.

UIManager has a new method, PauseGame. This is
subscribed to PlayerControl.TogglePauseUI.

This will enable or disable the PauseCanvas.
On enabled, it will set Time.timeScale to 0.
On disabled, it will set Time.timeScale to 1.

At the moment, there is an issue where the timer
coroutine is not paused.