Modify Timer and UIManager classes
Timer has a new delegate, CheckIsPaused.

Modified Countdown method in Timer class. An If
statement is included so that the timer
variable decrements, the timer text is updated,
and the TimeToSpawnNewCoin is called only when
delegate, CheckIsPaused, returns false.

UIManager has a new method, GetIsPaused. This
method returns _isPaused variable. This is
subscribed to Timer delegate, CheckIsPaused.

This fixes the issue where the timer text and
values were still being updated even when the
player is on the pause menu.