Modify UIs object
The ContinueButton in the PauseUI object in UIs object
is now set to call the PauseGame method on click.

There's an issue with the colour highlight being stuck
on "Selected Color" after clicking a button. Need to
find a workaround to set the button colour back to
"Normal Colour" after clicking a button. At the
moment, I've set the "Selected Color" to be the same
as the "Normal Color", as a temporary fix.