Modify UIManager, Timer, and ScoreTextProperties
Timer has a new delegate, TriggerEndGame. This is
invoked in Countdown method once the timer value
has reached zero.

UIManager has a new method, TriggerEndGame. This
is subscribed to Timer.TriggerEndGame.

UIManager has a new delegate, GetFinalScore. This
is invoked to get the score value from
ScoreTextProperties.

ScoreTextProperties subscribes GetCurrentScore to
UIManager.GetFinalScore.

Please note that the fetching and storing of
scores values stored in elements that are in
other canvases needs to occur before before
ToggleCanvase is called. Otherwise, a null
value is returned.