Modify PlayerControl
Modified Update loop so that player can only move
when UIManager._isPlaying is true. A new delegate,
CheckIsPlaying, is invoked that retrieves
_isPlaying.

UIManager has a new method, GetIsPlaying. This is
subscribed to UIManager.CheckIsPlaying

There is an issue where if the player is in the
trigger area of the coinhole and then replays the
game, _canDeposit remains true instead of being
reset to false.