Modify Timer and CoinSpawner
Timer has a new delegate, CheckCoinsDespawn. This
is invoked within Countdown method.

CoinSpawner has a new method, UpdateCoinDespawnTimers.
It is subscribed to Timer.CheckCoinsDespawn. The
intention when UpdateCoinDespawnTimers is called, it
will check all active and enabled coin prefabs, and
check their timer values. If the timer counter values
are above a certain threshold, the coin will be
despawned. Otherwise, their counter will be
incremented.