Modify CoinSpawner and CoinBehaviour
CoinBehaviour has a new method, UpdateCoinDespawnTimer.
This method evaluates whether DespawnTimerCurrent is less
than DespawnTimerMax. If true, the coin is disabled. If
false, DespawnTimerCurrent is incremented and the alpha
level of the sprite color will be decreased.

CoinSpawner.UpdateCoinDespawnTimers now looks through
the pooled object list for active coin prefabs i.e. coins
that are spawned in. If the coin prefab is active, then
it will call the prefab's
CoinBehaviour.UpdateCoinDespawnTimer method.